Retire untried tests
[Cherry-picked 2aa6fc6f951cc1407acca2d72930cbc68ad7ef46]
