--- tests/untried/neg/warn-unused-imports.check
@@ -1,33 +1,0 @@
-warn-unused-imports.scala:57: warning: Unused import
-    import p1.A // warn
-              ^
-warn-unused-imports.scala:62: warning: Unused import
-    import p1.{ A, B } // warn on A
-                ^
-warn-unused-imports.scala:67: warning: Unused import
-    import p1.{ A, B } // warn on both
-                ^
-warn-unused-imports.scala:67: warning: Unused import
-    import p1.{ A, B } // warn on both
-                   ^
-warn-unused-imports.scala:73: warning: Unused import
-    import c._  // warn
-             ^
-warn-unused-imports.scala:78: warning: Unused import
-    import p1._ // warn
-              ^
-warn-unused-imports.scala:85: warning: Unused import
-    import c._  // warn
-             ^
-warn-unused-imports.scala:91: warning: Unused import
-    import p1.c._  // warn
-                ^
-warn-unused-imports.scala:98: warning: Unused import
-    import p1._   // warn
-              ^
-warn-unused-imports.scala:118: warning: Unused import
-    import p1.A   // warn
-              ^
-error: No warnings can be incurred under -Xfatal-warnings.
-10 warnings found
-one error found
