--- tests/untried/neg/warn-unused-imports.flags
@@ -1,1 +1,0 @@
--Xfatal-warnings -Ywarn-unused-import
