--- tests/untried/neg/warn-unused-privates.flags
@@ -1,1 +1,0 @@
--Ywarn-unused -Xfatal-warnings
